requirements.txt
absl-py==2.0.0
attrs==23.1.0
contourpy==1.2.0
cycler==0.12.1
fonttools==4.45.1
kiwisolver==1.4.5
matplotlib==3.8.2
mediapipe==0.10.8
opencv-contrib-python==4.8.1.78
sounddevice==0.4.6
Brotli==1.0.9
MouseInfo==0.1.3
Pillow==10.1.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
archspec==0.2.1
boltons==23.0.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==2.0.4
chatgpt==2.2212.0
click==8.1.7
colorama==0.4.6
conda==23.10.0
conda-content-trust==0.2.0
conda-libmamba-solver==23.11.1
conda-package-handling==2.2.0
conda-package-streaming==0.9.0
cryptography==41.0.3
distro==1.8.0
elastic-transport==8.10.0
elasticsearch==8.11.0
flatbuffers==23.5.26
frozenlist==1.4.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.4
joblib==1.3.2
jsonpatch==1.32
jsonpointer==2.1
libmambapy==1.5.3
markdown-it-py==3.0.0
mdurl==0.1.2
menuinst==1.4.19
multidict==6.0.4
nltk==3.8.1
numpy==1.26.2
openai==1.3.7
opencv-python==4.8.1.78
packaging==23.1
pandas==2.1.3
pip==23.3
pluggy==1.0.0
protobuf==3.20.3
pyOpenSSL==23.2.0
pycosat==0.6.6
pycparser==2.21
pydantic==2.5.2
pydantic-core==2.14.5
pygments==2.17.2
pyparsing==3.1.1
pyperclip==1.8.2
python-dateutil==2.8.2
pytweening==1.0.7
pytz==2023.3.post1
regex==2023.10.3
requests==2.31.0
rich==13.7.0
ruamel.yaml==0.17.21
scipy==1.11.4
seaborn==0.11.2
setuptools==68.0.0
six==1.16.0
sniffio==1.3.0
tls-client==0.2.2
tqdm==4.65.0
truststore==0.8.0
typing-extensions==4.8.0
tzdata==2023.3
urllib3==1.26.18
wheel==0.41.2
win-inet-pton==1.1.0
yarl==1.9.3
zstandard==0.19.0
